codereview: 코리 반영해서 수정했습니다
